refactor: extract each cell component into own file
