+++ src/framework/theme/components/calendar/calendar-range.module.ts
@@ -7,11 +7,9 @@
 import { NgModule } from '@angular/core';
 
 import { NbCalendarRangeComponent } from './calendar-range.component';
-import {
-  NbCalendarRangeDayCellComponent,
-  NbCalendarRangeYearCellComponent,
-  NbCalendarRangeMonthCellComponent,
-} from './calendar-range-cells';
+import { NbCalendarRangeDayCellComponent } from './calendar-range-day-cell.component';
+import { NbCalendarRangeYearCellComponent } from './calendar-range-year-cell.component';
+import { NbCalendarRangeMonthCellComponent } from './calendar-range-month-cell.component';
 import { NbBaseCalendarModule } from './base-calendar.module';
